[PM-3905] chore: move webauthn utils to vault
