+++ apps/browser/src/vault/fido2/webauthn-utils.ts
